machine_learning/requirements.txt
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
sqlaclchemy==2.0.25
python-dotenv==1.0.0
shap==0.44.0